refactor: move prop types to another file
